requirements.txt
pdfplumber==0.10.3
pandas==2.1.3
requests==2.31.0
tenacity==8.2.3
ratelimit==2.2.1
geopandas==0.14.0
shapely==2.0.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
geojson==3.1.0
supabase==2.3.1
python-dotenv==1.0.0